+++ pages/_app.tsx
@@ -3,8 +3,8 @@ import { Provider } from 'react-redux'
 import { NextComponentType } from 'next'
 import App, { Container } from 'next/app'
 import withRedux from 'next-redux-wrapper'
-import initializeStore from '../src/states/store'
-import Auth from '../src/components/auth'
+import initializeStore from '../client/states/store'
+import Auth from '../client/components/auth'
 
 interface Props {
   Component: NextComponentType<any, any, any>
